+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for multiple collections indexing under the same parent.
 - Added code exclusions placeholder in the generation. (oneOf)
 - Added support for continuous access evaluation in Java. [#1179](https://github.com/microsoft/kiota/issues/1179)
+- Added support for special characters in URL query parameter names. [#1584](https://github.com/microsoft/kiota/pull/1584)
 
 ### Changed